requirements.txt

Django==4.2.25
# Phase 3 additions
reportlab==4.0.7  # For PDF generation
# Phase 4 additions
djangorestframework==3.15.0
flake8==6.1.0
coverage==7.2.7
